fix(integer-i64): widen execute_at_long + execute_at_text extras to i64
Follow-up to `edbc9f3`.  That commit caught the stale `as i32`
push in `execute_at_raw`, but two sibling worker entrypoints
(`execute_at_long` and `execute_at_text`) had identical stale
pushes.  No currently-failing test exercises them, but leaving
the asymmetry in place risks a silent post-2c slot-drift bug
surfacing later in text-returning or long-returning par()
workers.

Mechanical `put_stack(extra as i32)` → `put_stack(extra as i64)`
— two sites.  Tests: 16 → 16 (no observable change, no
regressions).  Preventive consistency fix.

+++ src/state/mod.rs
@@ -1782,7 +1782,7 @@ impl State {
         self.stack_pos = 4;
         self.put_stack(*arg);
         for &extra in extra_args {
-            self.put_stack(extra as i32);
+            self.put_stack(extra as i64);
         }
         self.put_stack(u32::MAX);
         self.code_pos = fn_pos;
@@ -1853,7 +1853,7 @@ impl State {
         self.stack_pos = 4;
         self.put_stack(*arg);
         for &extra in extra_args {
-            self.put_stack(extra as i32);
+            self.put_stack(extra as i64);
         }
         // Push the work buffer DbRefs as the hidden parameters.
         for cr in &work_crs {
